add haliax.debug.diagnose_common_problems
